+++ modules/tinymce/src/core/main/ts/api/EditorOptions.ts
@@ -5,7 +5,7 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-import { Arr, Obj, Strings, Type } from '@ephox/katamari';
+import { Obj, Strings, Type } from '@ephox/katamari';
 
 import Editor from './Editor';
 import { EditorOptions, NormalizedEditorOptions } from './OptionTypes';
@@ -207,14 +207,6 @@ const processDefaultValue = <T, U>(name: string, defaultValue: T, processor: Pro
 const create = (editor: Editor, initialOptions: Record<string, unknown>): Options => {
   const registry: Record<string, OptionSpec<any, any>> = {};
   const values: Record<string, any> = {};
-
-  editor.on('init', () => {
-    const unregisteredOptions = Arr.filter(Obj.keys(initialOptions), (option) => !isRegistered(option) && option !== 'mobile');
-    if (unregisteredOptions.length > 0) {
-      // eslint-disable-next-line no-console
-      console.warn('The following options were specified but have not been registered:\n - ' + unregisteredOptions.join('\n - '));
-    }
-  });
 
   const setValue = <T, U>(name: string, value: T, processor: SimpleProcessor | Processor<U>): boolean => {
     const result = processValue(value, processor);